+++ lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:im_stepper/src/gauge_stepper/circular_scrolling.dart';
+import 'package:im_stepper/src/gauge_stepper/gauge_stepper.dart';
 
 import 'stepper.dart';
 
@@ -12,13 +14,6 @@ class IconStepperDemo extends StatefulWidget {
 }
 
 class _IconStepperDemo extends State<IconStepperDemo> {
-  // MUST BE MAINTAINED, SEPARATELY.
-  int currentIndex = 0;
-
-  // THESE MUST BE USED TO CONTROL THE STEPPER FROM EXTERNALLY.
-  bool goNext = false;
-  bool goPrevious = false;
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -26,110 +21,11 @@ class _IconStepperDemo extends State<IconStepperDemo> {
         appBar: AppBar(
           title: Text('Icon Stepper Example'),
         ),
-        body: Column(
-          children: <Widget>[
-            Container(
-              decoration: BoxDecoration(
-                color: Colors.blueGrey,
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.grey,
-                    spreadRadius: 1.0,
-                    blurRadius: 2.0,
-                  )
-                ],
-                // borderRadius: BorderRadius.circular(5.0),
-              ),
-              child: IconStepper.externallyControlled(
-                goNext: goNext,
-                goPrevious: goPrevious,
-                direction: Axis.horizontal,
-                stepColor: Colors.white,
-                activeStepColor: Colors.amber,
-                lineColor: Colors.amberAccent,
-                lineLength: 75,
-                steppingEnabled: true,
-                icons: [
-                  Icon(Icons.home),
-                  Icon(Icons.person),
-                  Icon(Icons.account_balance),
-                  Icon(Icons.access_time),
-                ],
-              ),
-            ),
-            Container(
-              decoration: BoxDecoration(
-                border: Border.all(width: 0.1),
-              ),
-              padding: EdgeInsets.all(8.0),
-              alignment: Alignment.centerLeft,
-              child: Text(header()),
-            ),
-            Expanded(
-              child: Container(
-                margin: EdgeInsets.all(5.0),
-                child: FittedBox(
-                  child: Center(
-                    child: Text('${currentIndex + 1}'),
-                  ),
-                ),
-              ),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                RaisedButton(
-                  onPressed: () {
-                    // MUST TO CONTROL STEPPER FROM EXTERNAL BUTTONS.
-                    setState(() {
-                      goNext = false;
-                      goPrevious = true;
-
-                      if (currentIndex > 0) {
-                        currentIndex--;
-                      }
-                    });
-                  },
-                  child: Text('Previous'),
-                ),
-                RaisedButton(
-                  onPressed: () {
-                    // MUST TO CONTROL STEPPER FROM EXTERNAL BUTTONS.
-                    setState(() {
-                      goNext = true;
-                      goPrevious = false;
-
-                      if (currentIndex < 3) {
-                        currentIndex++;
-                      }
-                    });
-                  },
-                  child: Text('Next'),
-                ),
-              ],
-            )
-          ],
+        body: Padding(
+          padding: const EdgeInsets.all(0.0),
+          child: MyList(),
         ),
       ),
     );
-  }
-
-  String header() {
-    switch (currentIndex) {
-      case 0:
-        return 'Educational Background';
-
-      case 1:
-        return 'Professional Background';
-
-      case 2:
-        return 'Awards';
-
-      case 3:
-        return 'Sports';
-
-      default:
-        return 'Unknown';
-    }
   }
 }